feat: add support for windows using bash
This modifies the bin/lfe and bin/lfescript shell scripts to also work
on windows.

For bin/lfescript, all it does is adds another argument "-pa" to the
exec "$emulator" command with the ebin directory pointing at
"$LFE_ROOTDIR/ebin". This seems to fix lfescript on windows, while not
breaking it on macOS.

For bin/lfe, it adds an internal "emulator" variable which is set to the
"LFE_EMULATOR" environment variable, or erl if "LFE_EMULATOR" is not
present. This allows windows users to use "werl.exe" instead of
"erl.exe" as their emulator. Along with this, the script now checks for
the host operating system, and then adds some emulator options that
fixes starting lfe on windows. These fixes are not required when using
"werl.exe", so it only adds the fixes when the "LFE_EMULATOR" is not
"werl.exe". Then it launches the "$emulator" with "-pa
$LFE_ROOTDIR/ebin". This seems to fix lfe on windows, while not breaking
on macOS.

Running lfe requires a native bash shell on windows. This excludes WSL
for obvious reasons. I personally tested this with "4.4.23(1)-release
(x86_64-pc-msys)" which is shipped with windows' distribution of git.

Signed-off-by: Akshit Garg <garg.akshit@gmail.com>